# Improved performances for isPrime