The prime test performance has been improved compared to the previous version.