@josdejong 2019-10-02
Thanks Arthur! Does the algorithm you implemented have a specific name?

Can you please have a look at the failing tests "`n` is undefined? I also see a semicolon, that will let the linter break (`npm run lint`).