@arguiot 2019-10-02
Alright, I will try to fix that as soon as possible. I don’t know if it has a specific name, I use it in [TheoremJS](https://theorem.js.org) and I compared it with yours, and mine seems to be faster. It’s based on the multiple of 6 approach, described [here](http://mathforum.org/library/drmath/view/56068.html).