@josdejong 2019-10-03
@arguiot how can I easily test that your implementation is faster than the current one?