@ericman314 2019-10-03
Ooh, if we're running benchmarks, why not go big?

**Primes:**
524287
2147483647
87178291199
4398042316799
99194853094755497
2305843009213693951
11111111111111111111111

**Not primes:**
2199023255551
8796093022207
140737488355327
9007199254740991 (MAX_SAFE_INTEGER)
57646075230343487
147573952589676412927