@arguiot 2019-10-03
Why not, but we'll have to describe these numbers as BigInteger then.