@josdejong 2019-10-06
I've added a small benchmark: [test/benchmark/prime.js](https://github.com/josdejong/mathjs/blob/develop/test/benchmark/prime.js)

On my machine this gives the following results:

```
before improved algorithm of Arthur:

primes x 105 ops/sec ±0.41% (76 runs sampled)
not primes x 15,850 ops/sec ±0.30% (92 runs sampled)
carmichaels x 10,791,643 ops/sec ±0.23% (95 runs sampled)


after improved algorithm of Arthur:

primes x 158 ops/sec ±0.73% (80 runs sampled)
not primes x 23,573 ops/sec ±0.56% (94 runs sampled)
carmichaels x 10,595,211 ops/sec ±0.51% (96 runs sampled)
```

So that's a serious improvement :+1: . Thanks Arthur!