@josdejong 2021-10-14
Thanks for sharing @vitaly-t . Feel free to create a PR improving `isPrime` further 😄