@Yaffle 2021-10-14
the Miller-Rabin test is even faster if to implement it for small integers and if to run only after checking divisors <= 1000