@vitaly-t 2021-10-14
> the Miller-Rabin test is even faster if to implement it for small integers

I'm not sure about small numbers... [my implementation](https://github.com/vitaly-t/primes-generator/blob/main/src/is-prime.ts) can check the first million primes in just 217ms ([benchmarks](https://github.com/vitaly-t/primes-generator/tree/main/test)). How much faster Miller-Rabin can do it, though? From my tests, Miller-Rabin is good for larger numbers, while for small ones it is an overkill.