@arguiot 2021-10-14
I already implemented a Miller Rabin testing function in another language, but the code is pretty easy to read and reproduce in JS I think, so it's definitely translatable for `Number`.

I leave the source code here, I don't have the time right now to work on it:
https://github.com/arguiot/Euler/blob/master//Sources/Euler/NumberTheory/isPrime.swift#L88-L127